refactor: split uninitialisedRouter
